+++ TablePro/Models/Settings/MCPSettings.swift
@@ -8,6 +8,7 @@ struct MCPSettings: Codable, Equatable {
     var queryTimeoutSeconds: Int
     var logQueriesInHistory: Bool
     var requireAuthentication: Bool
+    var connectionApproval: MCPConnectionApproval
 
     static let `default` = MCPSettings(
         enabled: false,
@@ -16,7 +17,8 @@ struct MCPSettings: Codable, Equatable {
         maxRowLimit: 10_000,
         queryTimeoutSeconds: 30,
         logQueriesInHistory: true,
-        requireAuthentication: true
+        requireAuthentication: true,
+        connectionApproval: .oncePerConnection
     )
 
     init(
@@ -26,7 +28,8 @@ struct MCPSettings: Codable, Equatable {
         maxRowLimit: Int = 10_000,
         queryTimeoutSeconds: Int = 30,
         logQueriesInHistory: Bool = true,
-        requireAuthentication: Bool = true
+        requireAuthentication: Bool = true,
+        connectionApproval: MCPConnectionApproval = .oncePerConnection
     ) {
         self.enabled = enabled
         self.port = port
@@ -35,6 +38,7 @@ struct MCPSettings: Codable, Equatable {
         self.queryTimeoutSeconds = queryTimeoutSeconds
         self.logQueriesInHistory = logQueriesInHistory
         self.requireAuthentication = requireAuthentication
+        self.connectionApproval = connectionApproval
     }
 
     init(from decoder: Decoder) throws {
@@ -47,6 +51,9 @@ struct MCPSettings: Codable, Equatable {
         queryTimeoutSeconds = try container.decodeIfPresent(Int.self, forKey: .queryTimeoutSeconds) ?? 30
         logQueriesInHistory = try container.decodeIfPresent(Bool.self, forKey: .logQueriesInHistory) ?? true
         requireAuthentication = try container.decodeIfPresent(Bool.self, forKey: .requireAuthentication) ?? true
+        connectionApproval = try container.decodeIfPresent(
+            MCPConnectionApproval.self, forKey: .connectionApproval
+        ) ?? .oncePerConnection
 
         maxRowLimit = validatedMaxRowLimit
         defaultRowLimit = validatedDefaultRowLimit